feat: add lender and borrower pages with metadata handling
- Implemented LendPageClient for lenders to manage deposits and withdrawals, including real-time updates via SSE.
- Created LoansPageClient for borrowers to view their loans with pagination and status filtering.
- Developed LoanDetailsPageClient to display detailed loan information and repayment schedules.
- Introduced metadata utility to build dynamic page metadata based on locale and path.
- Enhanced layout with site URL configuration for better SEO and sharing capabilities.
